requirements.txt
flask
opencv-python
numpy
pillow
easyocr
sympy
python-dotenv
google-generativeai
reportlab
python-docx